Foundation: Flag types without padding bytes to optimize serialization
This allows some nice speedup for example for vectors/arrays of packed structs (Vector<Point2D> & co.) and in general structs without padding bytes that will get read and written in bulk.
Optimizations for now only on the non-versioned serializer.
I am not sure if it's worth making the versioned serializer more complicated trying to add this optimization.