Foundation: Improve StringIterator and StringView with better constexpr API design
This commit is doing a lot of work to allow better, easier and more error-proof parsing.
The goal is to always parse using StringIterator stepping and never directly manipulating pointer math.
StringIterator can only be constructed from a StringView. Also making many more StringView functions constexpr.
Add functions to check if a string is an integer or floating point number without parsing it.