Async: Add teardown phase to unify async state machine logic between all backends
An async that has been started, will go in the submission queue (Setup state).
If setup is successful, it will be activated.
After a (successful) activate an async is in Active state and it can be completed or cancelled.
An Active async that is completed, can be activated again (Submitting state) or it will be teardown.
Also when cancelling an Active async, a teardown will be issued to free resources created by setup.