Async: Remove private details from public header and generalize manual completion support
The removal of private details is mainly to reduce compile time (and LOC in general) for clients of the library.
The manual completion support has been generalized and shared between all backends.
Also we're separating a few completions into an executeOperation method to allow calling it from a thread pool task in the (hopefully) near future.
As the AsyncResult is not copyable, completion data (outputs / results from completions) are now saved into a dedicated CompletionData structures.